Reunifies the two divergent lines (merge-base 45349d9):

From chessckers-port (Gumbel Stage-1, 79702a7):
- improved_policy Gumbel target softmax(logP + sigma(completedQ)) computed at
  the searched root and emitted per cc::PureRecord (chunk.hpp/selfplay.hpp)

From ctl/pre-gumbel-run16:
- league opponent sampling: per-game past-champion pool for player2 (4332013)
- --league-probs PFSP-weighted league opponent sampling (709d060)
- blacks_move derived from position, not per-ply toggle (ee64b19)
- thread-safe ChesskersComputation::AddInput via AtomicVector (14fdc61)
- uint8 num_edges_ overflow fix for >255-move positions (f8e19d1)
- per-node scratch buffers widened to kMaxNumEdges=1024 (fec5291)
- quiet diagonal stops at blocking White piece (46d3ac1)
- run-14 kStartposFen comment update (7063569)

Only src/selfplay/game.cc was touched by both sides; auto-merge verified to
keep both the Gumbel emission block and the position-derived blacks_move.

+++ src/search/dag_classic/node.cc
@@ -266,7 +266,7 @@ void LowNode::SetBounds(GameResult lower, GameResult upper) {
   upper_bound_ = upper;
 }
 
-uint8_t Node::GetNumEdges() const {
+uint16_t Node::GetNumEdges() const {
   return low_node_ ? low_node_->GetNumEdges() : 0;
 }
